pyproject.toml
[tool.poetry]
name = "munros"
version = "0.1.0"
description = ""
authors = ["corin <corinlwilkins@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
OSGridConverter = "^0.1.3"
click = "^8.1.7"
ortools = "^9.7.2996"
plotly-express = "^0.4.1"
bng-latlon = "^1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
parse = "munros.main:parse_csv"
deltas = "munros.main:deltas"
main = "munros.main:main"
plot = "munros.plot:main"